+++ app/fbcnms-projects/inventory/app/components/assurance/KqiSourceFormEdit.js
@@ -14,10 +14,8 @@ import EditKqiSourceMutation from '../../mutations/EditKqiSourceMutation';
 import React from 'react';
 import fbt from 'fbt';
 
-import Button from '@material-ui/core/Button';
+import ButtonSaveDelete from './common/ButtonSaveDelete';
 import Card from '@symphony/design-system/components/Card/Card';
-import ConfigureTitle from './common/ConfigureTitle';
-import FormField from '@symphony/design-system/components/FormField/FormField';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
 import {useDisabledButtonEdit} from './common/useDisabledButton';
@@ -69,11 +67,6 @@ const useStyles = makeStyles(() => ({
   },
   header: {
     marginBottom: '1rem',
-  },
-  option: {
-    width: '111px',
-    height: '36px',
-    alignSelf: 'flex-end',
   },
   delete: {
     display: 'flex',
@@ -159,40 +152,30 @@ const KqiSourceFormEdit = (props: Props) => {
         direction="row"
         justify="flex-end"
         alignItems="center">
-        <Grid>
+        <Grid item xs>
           <Text variant="h6" weight={'bold'}>
             {fbt('Edit KQI Source', ' ')}
           </Text>
         </Grid>
-        <Grid item xs>
-          <FormField>
-            <Button
-              style={{marginRight: '1rem'}}
-              className={classes.option}
-              variant="outlined"
-              color="primary"
-              onClick={() => hideKqiSourceFormEdit()}>
-              Cancel
-            </Button>
-          </FormField>
+        <Grid>
+          <ButtonSaveDelete
+            variant="outlined"
+            onClick={() => hideKqiSourceFormEdit()}>
+            Cancel
+          </ButtonSaveDelete>
         </Grid>
         <Grid>
-          <FormField>
-            <Button
-              className={classes.option}
-              variant="contained"
-              color="primary"
-              onClick={() => {
-                hideKqiSourceFormEdit();
-                handleClick();
-              }}
-              disabled={handleDisable}>
-              Save
-            </Button>
-          </FormField>
+          <ButtonSaveDelete
+            onClick={() => {
+              hideKqiSourceFormEdit();
+              handleClick();
+            }}
+            disabled={handleDisable}>
+            Save
+          </ButtonSaveDelete>
         </Grid>
       </Grid>
-      <Grid className={classes.container} item xs>
+      <Grid item xs>
         <Card>
           <Grid container className={classes.formField} spacing={2}>
             <Grid item xs={12} lg={6}>