+++ app/fbcnms-projects/inventory/app/components/work_orders/SelectAvailabilityAssignee.js
@@ -111,9 +111,9 @@ const SelectAvailabilityAssignee = (props: Props) => {
     setAppointment(false);
   }, []);
 
-  const setInfo = (label: string, user: ShortUser) => {
+  const setInfo = (label: string, user: ShortUser, startDate: string) => {
     _setWorkOrderDetail(label, user);
-    setAppointment(true);
+    setAppointment(true, startDate);
   };
 
   const applyFilters = () => {
@@ -125,10 +125,13 @@ const SelectAvailabilityAssignee = (props: Props) => {
     setAppointment(false);
   };
 
-  const setAppointment = saveAppointment => {
+  const setAppointment = (
+    saveAppointment,
+    selectedDate = slotStartDate.toISOString(),
+  ) => {
     setAppointmentData({
       duration,
-      date: slotStartDate.toISOString(),
+      date: selectedDate,
       saveAppointment: saveAppointment,
     });
   };
@@ -164,6 +167,7 @@ const SelectAvailabilityAssignee = (props: Props) => {
                 <DateTimePicker
                   variant="inline"
                   inputVariant="outlined"
+                  minDate={moment()}
                   value={slotStartDate}
                   onChange={setSlotStartDate}
                   onClose={orderDatesValidation}
@@ -261,7 +265,9 @@ const SelectAvailabilityAssignee = (props: Props) => {
         {useFilters ? (
           <UserByAppointmentTypeahead
             selectedUser={workOrder.assignedTo || undefined}
-            onUserSelection={user => setInfo('assignedTo', user)}
+            onUserSelection={(user, startDate) =>
+              setInfo('assignedTo', user, startDate)
+            }
             slotStartDate={slotStartDate.toISOString()}
             slotEndDate={slotEndDate.toISOString()}
             duration={duration}