+++ graph/resolverutil/resource_specification_filters.go
@@ -8,6 +8,7 @@ import (
 	"github.com/facebookincubator/symphony/graph/graphql/models"
 	"github.com/facebookincubator/symphony/pkg/ent"
 	"github.com/facebookincubator/symphony/pkg/ent/resourcespecification"
+	"github.com/facebookincubator/symphony/pkg/ent/resourcetype"
 	"github.com/facebookincubator/symphony/pkg/ent/schema/enum"
 	"github.com/pkg/errors"
 )
@@ -16,7 +17,8 @@ func handleResourceSpecificationFilter(q *ent.ResourceSpecificationQuery, filter
 	switch filter.FilterType {
 	case models.ResourceSpecificationFilterTypeName:
 		return resourceSpecificationNameFilter(q, filter)
-
+	case models.ResourceSpecificationFilterTypeResourceType:
+		return resourceSpecificationNameFilter(q, filter)
 	default:
 		return nil, errors.Errorf("filter type is not supported: %s", filter.FilterType)
 	}
@@ -28,6 +30,13 @@ func resourceSpecificationNameFilter(q *ent.ResourceSpecificationQuery, filter *
 		return q.Where(resourcespecification.NameContainsFold(*filter.StringValue)), nil
 	} else if filter.Operator == enum.FilterOperatorIs && filter.StringValue != nil {
 		return q.Where(resourcespecification.NameEQ(*filter.StringValue)), nil
+	}
+	return nil, errors.Errorf("operation is not supported: %s", filter.Operator)
+}
+
+func resourceSpecificationResourceTypeFilter(q *ent.ResourceSpecificationQuery, filter *models.ResourceSpecificationFilterInput) (*ent.ResourceSpecificationQuery, error) {
+	if filter.Operator == enum.FilterOperatorIs && filter.IDSet != nil {
+		return q.Where(resourcespecification.HasResourcetypeWith(resourcetype.IDIn(filter.IDSet...))), nil
 	}
 	return nil, errors.Errorf("operation is not supported: %s", filter.Operator)
 }